Fix lgtm alerts issues
