+++ device/accton/x86_64-accton_csp7551-r0/sonic_platform/eeprom.py
@@ -1,5 +1,4 @@
 try:
-    import os
     import sys
     import re
     if sys.version_info[0] >= 3:
