+++ device/accton/x86_64-accton_csp7551-r0/sonic_platform/psu.py
@@ -68,7 +68,7 @@ class Psu(PsuBase):
         command=IPMI_SENSOR_PARAM_CMD.format(PSU_SENSOR_INDEX[self.index][type])
         st1, log1 = self._syscmd(command)
         if st1 != 0:
-            raise 'error on syscmd'
+            print('error on syscmd')
         m1 = log1.split()[2].strip()
         m2 = log1.split()[3].strip()
         m2 = (int(m2,16)&0xc0) <<2
